#5-1) company structure refactor
2026/04/17-refactor/company
